login_backend/requirements.txt
fastapi==0.115.12
passlib==1.7.4
pydantic==2.11.2
python-dotenv==1.1.0
SQLAlchemy==2.0.40
